src/core/bot/nn/requirements-train.txt
torch>=2.0
python-chess>=1.999
tqdm>=4.66
